feat: add x line and remove border
